# swagger/src/main/resources/swagger-cost-seiko.yaml
openapi: 3.0.3
info:
  title: Cost seiko
  description: |-
    Cost Seiko API
  contact:
    email: khalil@lokalix-solutions.com
  version: 1.0.0
servers:
  - url: http://localhost:9090
...
paths:
...
  /units/search:
    post:
      tags:
        - unit
      summary: |
        Return paginated data ordered alphabetically
      operationId: searchUnits
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitsPaginated"
  /units/{uid}:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - unit
      summary: Retrieve a unit
      operationId: retrieveUnit
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
    put:
      tags:
        - unit
      summary: Update a unit
      operationId: updateUnit
      requestBody:
        description: update unit
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitUpdate'
        required: true
      responses:
        '200':
          description: object updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unit'
  /units:
    get:
      tags:
        - unit
      summary: List all units
      operationId: listAllUnits
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Unit'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - unit
      summary: Create a new unit
      operationId: createUnit
      requestBody:
        description: New unit
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnitCreate'
        required: true
      responses:
        '201':
          description: object created
          headers:
            Uid:
              schema:
                $ref: '#/components/schemas/Uid'
              description: uid of the created object.
  /units/{uid}/archive:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - unit
      summary: Archive a unit
      operationId: archiveUnit
      responses:
        '200':
          description: successful operation
  /units/exists-by-name:
    post:
      tags:
        - unit
      summary: check if exists by name
      operationId: existUnitByName
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StringBody'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Uid'
...
  /cost-requests/{uid}/lines/{lineUid}/toolings:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/lineUidParam'
    post:
      tags:
        - cost-request-line
      summary: create a tooling cost line
      operationId: createToolingCostLine
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolingCostLineCreate"
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolingCostLine'
...
  /cost-requests/{uid}/lines/{lineUid}/toolings/{costingLineUid}/files/download:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/lineUidParam'
      - $ref: '#/components/parameters/costingLineUidParam'
    post:
      tags:
        - cost-request-line
      summary: download tooling cost line files
      operationId: downloadToolingCostLineFiles
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Uid"
      responses:
        '200':
          description: Successful operation
...
  /toolings/to-be-estimated/search:
    post:
      tags:
        - tooling
      summary: |
        Search toolings to be estimated with filter
        Return paginated data ordered by creationDate DESC
      operationId: searchToolingsToBeEstimated
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolingCostLinesPaginated"
  /toolings/to-be-estimated/{uid}/files:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - tooling
      summary: get to be estimated tooling cost line files metadata
      operationId: getToBeEstimatedToolingCostLineFilesMetadata
      parameters:
        - $ref: '#/components/parameters/uidParam'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileInfo'
  /toolings/to-be-estimated/{uid}/files/download:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - tooling
      summary: download to be estimated tooling cost line files
      operationId: downloadToBeEstimatedToolingCostLineFiles
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Uid"
      responses:
        '200':
          description: Successful operation
  /toolings/to-be-estimated/{uid}/estimate:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - tooling
      summary: estimate a tooling cost line
      operationId: estimateToolingCostLine
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolingCostLineEstimate"
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolingCostLine'
  /toolings/to-be-estimated/{uid}/reject:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - tooling
      summary: reject a tooling cost line
      operationId: rejectToolingCostLine
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolingCostLineReject"
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolingCostLine'
  /toolings/to-be-estimated/{uid}/messages:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - tooling
      summary: get tooling messages
      operationId: getAllToolingMessages
      responses:
        '200':
          description: successful operation, all tooling messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      tags:
        - tooling
      summary: create tooling message
      operationId: createToolingMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /toolings/to-be-estimated/{uid}/messages/{messageUid}:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/messageUidParam'
    put:
      tags:
        - tooling
      summary: update tooling message
      operationId: updateToolingMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageUpdate'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /toolings/to-be-estimated/{uid}/messages/{messageUid}/delete:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/messageUidParam'
    post:
      tags:
        - tooling
      summary: delete tooling message
      operationId: deleteToolingMessage
      responses:
        '200':
          description: Successful operation, cost request line message is deleted
  /toolings/to-be-estimated/{uid}/messages/{messageUid}/undo:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/messageUidParam'
    post:
      tags:
        - tooling
      summary: undo deleted tooling message
      operationId: undoToolingMessage
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
components:
  parameters:
    uidParam:
      in: path
      name: uid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
    offsetParam:
      in: query
      name: offset
      required: true
      schema:
        type: integer
        minimum: 0
        default: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: The numbers of items to return.
...
    lineUidParam:
      in: path
      name: lineUid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
      description: The line UID
    costingLineUidParam:
      in: path
      name: costingLineUid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
      description: The costing line UID
...
    messageUidParam:
      in: path
      name: messageUid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
...
  schemas:
    StringBody:
      type: object
      required:
        - value
      properties:
        value:
          type: string
...
    ToolingCostLinesPaginated:
      type: object
      allOf:
        - $ref: "#/components/schemas/PaginationInfos"
        - type: object
          required:
            - results
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/ToolingCostLine"
    ToolingCostLine:
      allOf:
        - $ref: "#/components/schemas/ToolingCostLineCreate"
        - type: object
          required:
            - uid
            - nbFiles
          properties:
            uid:
              $ref: '#/components/schemas/Uid'
            totalCostInCurrency:
              type: number
              format: double
            outsourced:
              type: boolean
            outsourcingStatus:
              $ref: '#/components/schemas/OutsourcingStatus'
            rejectReason:
              type: string
            nbFiles:
              type: integer
            nbMessages:
              type: integer
    ToolingCostLineCreate:
      type: object
      required:
        - description
        - quantity
      properties:
        description:
          type: string
        quantity:
          type: number
          format: double
        unitCostInCurrency:
          type: number
          format: double
          description: not used if tooling is outsourced
        toolingPartNumber:
          type: string
...
    ToolingCostLineEstimate:
      type: object
      required:
        - unitCostInCurrency
      properties:
        unitCostInCurrency:
          type: number
          format: double
    ToolingCostLineReject:
      type: object
      required:
        - rejectReason
      properties:
        rejectReason:
          type: string
...
    PaginationInfos:
      type: object
      required:
        - total
        - page
        - per_page
        - has_next
        - has_prev
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
    BasicSearch:
      type: object
      required:
        - searchText
      properties:
        searchText:
          type: string
...
    FileInfo:
      type: object
      required:
        - uid
        - fileName
        - type
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        fileName:
          type: string
        type:
          $ref: '#/components/schemas/FileType'
...
    Uid:
      type: string
      format: uuid
...
    FileType:
      type: string
      enum:
        - ANY
        - QUOTATION_PDF
...
    Message:
      type: object
      allOf:
        - $ref: '#/components/schemas/MessageCreate'
        - type: object
          required:
            - uid
            - creationDate
            - lastModificationDate
            - userUid
            - userLogin
            - deleted
            - updated
          properties:
            uid:
              $ref: '#/components/schemas/Uid'
            creationDate:
              type: string
            lastModificationDate:
              type: string
            userUid:
              $ref: '#/components/schemas/Uid'
            userLogin:
              type: string
            deleted:
              type: boolean
            updated:
              type: boolean
    MessageCreate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
    MessageUpdate:
      type: object
      required:
        - content
      properties:
        content:
          type: string
...
    UnitsPaginated:
      type: object
      allOf:
        - $ref: "#/components/schemas/PaginationInfos"
        - type: object
          required:
            - results
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/Unit"
    UnitCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    UnitUpdate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    Unit:
      type: object
      required:
        - uid
        - name
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        name:
          type: string
...
    OutsourcingStatus:
      type: string
      enum:
        - TO_BE_ESTIMATED
        - ESTIMATED
        - REJECTED
...
    CustomError:
      type: object
      required:
        - code
      properties:
        code:
          $ref: '#/components/schemas/CustomErrorCode'
        message:
          type: string
    CustomErrorCode:
      type: string
      enum:
        - ACCESS_FORBIDDEN
        - INTERNAL_ERROR
        - INVALID_EMAIL_OR_PASSWORD
        - ENTITY_NOT_FOUND
        - ENTITY_ALREADY_EXISTS
        - LICENSE_LIMIT_EXCEEDED
        - USER_IS_INACTIVE
        - USER_RESET_PASSWORD_INVALID_TOKEN
        - USER_CANNOT_DEACTIVATE_ONESELF
        - FILE_UNSUPPORTED
        - FILE_EXISTS
        - FILE_NOT_FOUND
        - GENERIC_ERROR
        - MUST_BE_DISCONNECTED
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CustomError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

# 2) Apply the security globally to all operations
security:
  - bearerAuth: [ ]         # use the same name as above